refactor search into hook and split mobile desktop
